+++ WebEngineeringFrontend/my-react-app/src/pages/popGdp/PopGdp.js
@@ -1,31 +1,25 @@
 import { useState } from "react";
 import "./PopGdp.css";
 import NavBar from "../../components/NavBar";
-import ResultsHeader from "../../components/ResultsHeader";
-import Results from "../../components/Results";
 import WebFormSmall from "../../components/WebFormSmall";
 import WebFormBig from "../../components/WebFormBig";
 
 function PopGdp() {
-    const [dataRetrieve, setDataRetrieve] = useState([]);
     const [countryRetrieve, setCountryRetrieve] = useState("");
-    const [yearRetrieve, setYearRetrieve] = useState();
+    const [yearRetrieve, setYearRetrieve] = useState("");
 
-    const [dataCreate, setDataCreate] = useState([]);
     const [countryCreate, setCountryCreate] = useState("");
-    const [yearCreate, setYearCreate] = useState();
-    const [populationCreate, setPopulationCreate] = useState();
-    const [gdpCreate, setGdpCreate] = useState();
+    const [yearCreate, setYearCreate] = useState("");
+    const [populationCreate, setPopulationCreate] = useState("");
+    const [gdpCreate, setGdpCreate] = useState("");
 
-    const [dataUpdate, setDataUpdate] = useState([]);
     const [countryUpdate, setCountryUpdate] = useState("");
-    const [yearUpdate, setYearUpdate] = useState();
-    const [populationUpdate, setPopulationUpdate] = useState();
-    const [gdpUpdate, setGdpUpdate] = useState();
+    const [yearUpdate, setYearUpdate] = useState("");
+    const [populationUpdate, setPopulationUpdate] = useState("");
+    const [gdpUpdate, setGdpUpdate] = useState("");
 
-    const [dataDelete, setDataDelete] = useState([]);
     const [countryDelete, setCountryDelete] = useState("");
-    const [yearDelete, setYearDelete] = useState();
+    const [yearDelete, setYearDelete] = useState("");
 
     const changeCountryRetrieve = event => {
         setCountryRetrieve(event.target.value);
@@ -77,18 +71,17 @@ function PopGdp() {
 
     const handleSubmitRetrieve = (event) => {
         event.preventDefault();
-        setDataCreate([]);
-        setDataUpdate([]);
-        setDataDelete([]);
-        if (countryRetrieve.length > 3) {
+        if (countryRetrieve.length > 3 && yearRetrieve.length > 0) {
             fetch("http://localhost:8080/countries/" + countryRetrieve + "/" + yearRetrieve + "/population-gdp-name")
             .then(response => response.json())
-            .then(json => setDataRetrieve(json))
+            .then(json => localStorage.setItem("results", JSON.stringify(json)))
+            .then(() => window.location.href = "results")
             .catch(error => alert("Data not found"));
-        } else if (countryRetrieve.length === 3) {
+        } else if (countryRetrieve.length === 3 && yearRetrieve.length > 0) {
             fetch("http://localhost:8080/countries/" + countryRetrieve.toUpperCase() + "/" + yearRetrieve + "/population-gdp-code")
             .then(response => response.json())
-            .then(json => setDataRetrieve(json))
+            .then(json => localStorage.setItem("results", JSON.stringify(json)))
+            .then(() => window.location.href = "results")
             .catch(error => alert("Data not found"));
         } else {
             alert("Something is wrong");
@@ -97,10 +90,7 @@ function PopGdp() {
 
     const handleSubmitCreate = (event) => {
         event.preventDefault();
-        setDataRetrieve([]);
-        setDataUpdate([]);
-        setDataDelete([]);
-        if (countryCreate.length > 3) {
+        if (countryCreate.length > 3 && yearCreate.length > 0) {
             const requestOptions = {
                 method: 'POST',
                 headers: { 'Content-Type': 'application/json' },
@@ -108,9 +98,10 @@ function PopGdp() {
             };
             fetch("http://localhost:8080/countries/country/population-gdp-name", requestOptions)
             .then(response => response.json())
-            .then(json => setDataCreate(json))
+            .then(json => localStorage.setItem("results", JSON.stringify(json)))
+            .then(() => window.location.href = "results")
             .catch(error => alert("Data not found"));
-        } else if (countryCreate.length === 3) {
+        } else if (countryCreate.length === 3 && yearCreate.length > 0) {
             const requestOptions = {
                 method: 'POST',
                 headers: { 'Content-Type': 'application/json' },
@@ -118,19 +109,17 @@ function PopGdp() {
             };
             fetch("http://localhost:8080/countries/iso-code/population-gdp-code", requestOptions)
             .then(response => response.json())
-            .then(json => setDataCreate(json))
+            .then(json => localStorage.setItem("results", JSON.stringify(json)))
+            .then(() => window.location.href = "results")
             .catch(error => alert("Data not found"));
         } else {
-            alert("We do not support such format yet");
+            alert("Something is wrong");
         }
     }
 
     const handleSubmitUpdate = (event) => {
         event.preventDefault();
-        setDataCreate([]);
-        setDataRetrieve([]);
-        setDataDelete([]);
-        if (countryUpdate.length > 3) {
+        if (countryUpdate.length > 3 && yearUpdate.length > 0) {
             const requestOptions = {
                 method: 'PUT',
                 headers: { 'Content-Type': 'application/json' },
@@ -138,9 +127,10 @@ function PopGdp() {
             };
             fetch("http://localhost:8080/countries/" + countryUpdate + "/" + yearUpdate + "/population-gdp-name", requestOptions)
             .then(response => response.json())
-            .then(json => setDataUpdate(json))
+            .then(json => localStorage.setItem("results", JSON.stringify(json)))
+            .then(() => window.location.href = "results")
             .catch(error => alert("Data not found"));
-        } else if (countryUpdate.length === 3) {
+        } else if (countryUpdate.length === 3 && yearUpdate.length > 0) {
             const requestOptions = {
                 method: 'PUT',
                 headers: { 'Content-Type': 'application/json' },
@@ -148,35 +138,35 @@ function PopGdp() {
             };
             fetch("http://localhost:8080/countries/" + countryUpdate.toUpperCase() + "/" + yearUpdate + "/population-gdp-code", requestOptions)
             .then(response => response.json())
-            .then(json => setDataUpdate(json))
+            .then(json => localStorage.setItem("results", JSON.stringify(json)))
+            .then(() => window.location.href = "results")
             .catch(error => alert("Data not found"));
         } else {
-            alert("We do not support such format yet");
+            alert("Something is wrong");
         }
     }
 
     const handleSubmitDelete = (event) => {
         event.preventDefault();
-        setDataCreate([]);
-        setDataUpdate([]);
-        setDataRetrieve([]);
-        if (countryDelete.length > 3) {
+        if (countryDelete.length > 3 && yearDelete.length > 0) {
             const requestOptions = {
                 method: 'DELETE',
                 headers: { 'Content-Type': 'application/json' },
             };
             fetch("http://localhost:8080/countries/" + countryDelete + "/" + yearDelete + "/" + "population-gdp-name", requestOptions)
             .then(response => response.json())
-            .then(json => setDataDelete(json))
+            .then(json => localStorage.setItem("results", JSON.stringify(json)))
+            .then(() => window.location.href = "results")
             .catch(error => alert("Data not found"));
-        } else if (countryDelete.length === 3) {
+        } else if (countryDelete.length === 3 && yearDelete.length > 0) {
             const requestOptions = {
                 method: 'DELETE',
                 headers: { 'Content-Type': 'application/json' },
             };
             fetch("http://localhost:8080/countries/" + countryDelete.toUpperCase() + "/" + yearDelete + "/" + "population-gdp-code", requestOptions)
             .then(response => response.json())
-            .then(json => setDataDelete(json))
+            .then(json => localStorage.setItem("results", JSON.stringify(json)))
+            .then(() => window.location.href = "results")
             .catch(error => alert("Data not found"));
         } else {
             alert("Something is wrong");
@@ -195,19 +185,6 @@ function PopGdp() {
                 <WebFormBig header="Update" handleSubmit={handleSubmitUpdate} country={countryUpdate} population={populationUpdate} gdp={gdpUpdate} year={yearUpdate} changeCountry={changeCountryUpdate} changePopulation={changePopulationUpdate} changeGdp={changeGdpUpdate} changeYear={changeYearUpdate} button="Update"/>
                 <WebFormSmall header="Delete" handleSubmit={handleSubmitDelete} country={countryDelete} year={yearDelete} changeCountry={changeCountryDelete} changeYear={changeYearDelete} button="Delete"/>
             </div>
-            <ul className="results-list">
-                <ResultsHeader empty={dataRetrieve.length === 0 && dataCreate.length === 0 && dataUpdate.length === 0 && dataDelete.length === 0}/>
-                <Results retrieve={dataRetrieve.length != 0}
-                create={dataCreate.length != 0}
-                update={dataUpdate.length != 0}
-                delete={dataDelete.length != 0}
-                populationRetrieve={dataRetrieve.population}
-                gdpRetrieve={dataRetrieve.gdp}
-                id={dataCreate.id}
-                populationUpdate={dataUpdate.population}
-                gdpUpdate={dataUpdate.gdp}
-                message={dataDelete.message} />
-            </ul> 
         </div>
     );
 }
